Refactor JobPosts controller, add route under /jobs
